\name{adworld}
\docType{data}
\alias{adworld}
\encoding{latin1}
\title{Geographical coordinates}
\description{Latitude and longitude of all administrative areas.}
\usage{data(adworld)}
\format{A matrix of many rows and 3 columns (Latitude, Longitude and name of the administrative area)}
\source{Latitude and longitude coordinates of the administrative areas were obtained from the web page \url{https://www.openstreetmap.org}.}
\keyword{adworld}
