\name{Beetles}
\docType{data}
\alias{Beetles}
\title{Individual counts of species of beetles}
\description{This database includes 15,142 records belonging to 54 Iberian species of the Scarabaeidae (Coleoptera) previously compiled in the so called BANDASCA database (Lobo & Martín-Piera, 2002) also freely available in GBIF (\url{http://www.gbif.org/}). Individual counts, longitude and latitude of species occurrences of the family Scarabaeidae in the Iberian Peninsula are provided.}
\usage{data(Beetles)}
\format{A matrix with four columns: species, longitude, latitude and individual counts.}
\references{
Lobo, J.M. & Martín-Piera, F. 2002. Searching for a predictive model for Iberian dung beetle species richness based on spatial and environmental variables. \emph{Conservation Biology} 16: 158-173.
}
\keyword{Beetles}

 
