\name{Estimators}
\docType{data}
\alias{Estimators}
\encoding{latin1}
\title{Estimators obtained with the function KnowBPolygon}
\usage{data(Estimators)}
\description{Estimators obtained with the function \link[KnowBR]{KnowBPolygon} using species of freshwater fish species in all the countries of the world (Guisande et al., 2010).} 
\references{
Guisande, C., Manjarrés-Hernández, A., Pelayo-Villamil, P., Granado-Lorencio, C., Riveiro, I., Acuña, A., Prieto-Piraquive, E., Janeiro, E., Matías, J.M., Patti, C., Patti, B., Mazzola, S., Jiménez, L.F., Duque, S. & Salmerón, F. (2010) IPez: An expert system for the taxonomic identification of fishes based on machine learning techniques. \emph{Fisheries Research}, 102, 240-247.}
\keyword{Estimators}
