\name{FishIrelandUK}
\docType{data}
\alias{FishIrelandUK}
\title{Estimators obtained with the function KnowB}
\usage{data(FishIrelandUK)}
\description{Estimators obtained with the function \link[KnowBR]{KnowB} using a database that includes 121,709 records of freshwater fishes obtained from Pelayo-Villamil et al. (2015), with the Clench estimator and a cell resolution of 5'.}
\references{
Pelayo-Villamil, P., Guisande, C., Vari, R.P., Manjarrés-Hernández, A., García-Roselló, E., González-Dacosta, J., Heine, J., González-Vilas, L., Patti, B., Quinci, E.M., Jiménez, L.F., Granado-Lorencio, C., Tedesco, P.A., Lobo, J.M. (2015) Global diversity patterns of freshwater fishes-Potential victims of their own success. \emph{Diversity and Distributions}, 21: 345-356.
}
\keyword{FishIrelandUK}

 
